Renaming all files and routes to dashed
