+++ Source/VolunteerReporting/Web.Angular/src/app/app.component.html
@@ -9,7 +9,7 @@
 
     <div class="container">
         <nav>
-            <a routerLink="/casereport/list" routerLinkActive="active">Case Report List</a>
+            <a routerLink="/case-report/list" routerLinkActive="active">Case Report List</a>
         </nav>
         <router-outlet></router-outlet>
     </div>
