+++ Source/VolunteerReporting/Web.Angular/src/app/case-report/case-report.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 
-import { CaseReportRouting } from './casereport-routing.module';
-import { CaseReportListComponent } from './casereportlist/casereportlist.component';
+import { CaseReportRouting } from './case-report-routing.module';
+import { CaseReportListComponent } from './case-report-list/case-report-list.component';
 
 @NgModule({
     imports: [
