+++ Source/VolunteerReporting/Web.Angular/src/app/shared/models/index.ts
@@ -1,3 +1,3 @@
 export * from './location.model';
-export * from './casereport.model';
-export * from './anonymouscasereport.model';
+export * from './case-report.model';
+export * from './anonymous-case-report.model';
